Update Namesilo.php
I removed throwing exception when someone is not using Payment ID, this prevented them for using module without Payment ID defined.
But there is an issue, this is how input box is showing. 

<input id="el-Payment_ID" type="Payment_ID" name="config[Payment_ID]" value="" required="required">

It is set to REQUIRED and I have no idea how do i change that... I'm going to contact boxbilling musicboxlimited for this and i'm gonna do another pull request.
So basically, once you define Payment ID only way to remove it is to inspect element and remove required="required" and then save it.